% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_health_data.R
\name{clean_health_data}
\alias{clean_health_data}
\title{Clean health mortality data (customized for CDC dataset)}
\usage{
clean_health_data(df)
}
\arguments{
\item{df}{Raw CDC mortality dataset.}
}
\value{
A cleaned tibble with standardized columns:
date, year, month, heart, respiratory, covid
}
\description{
Clean health mortality data (customized for CDC dataset)
}
