% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trends.R
\name{plot_trends}
\alias{plot_trends}
\title{Plot mortality trends for a single cause}
\usage{
plot_trends(df, cause = c("heart", "respiratory", "covid"))
}
\arguments{
\item{df}{A cleaned dataset created by clean_health_data().}

\item{cause}{Which cause to plot: "heart", "respiratory", or "covid".}
}
\value{
A ggplot2 line plot.
}
\description{
Plot mortality trends for a single cause
}
