% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trends.R
\name{plot_trends_all}
\alias{plot_trends_all}
\title{Plot mortality trends for all three causes}
\usage{
plot_trends_all(df)
}
\arguments{
\item{df}{A cleaned dataset created by clean_health_data().}
}
\value{
A ggplot2 line plot.
}
\description{
Creates a combined line plot with heart, respiratory,
and covid deaths over time.
}
