% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.health_summary}
\alias{print.health_summary}
\title{Print method for health_summary objects}
\usage{
\method{print}{health_summary}(x, ...)
}
\arguments{
\item{x}{An object of class "health_summary".}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for health_summary objects
}
