% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_causes.R
\name{summarize_causes}
\alias{summarize_causes}
\title{Summarize yearly mortality by cause}
\usage{
summarize_causes(df)
}
\arguments{
\item{df}{A cleaned dataset created by clean_health_data().}
}
\value{
An object of class "health_summary" containing:
\itemize{
\item data: a tibble with yearly totals
\item n_years: number of years
\item causes: list of included causes
}
}
\description{
Summarize yearly mortality by cause
}
