% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{summary.health_summary}
\alias{summary.health_summary}
\title{Summary method for health_summary objects}
\usage{
\method{summary}{health_summary}(object, ...)
}
\arguments{
\item{object}{An object of class "health_summary".}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for health_summary objects
}
